requirements.txt
apturl==0.5.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.5
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==3.2.18
django-extensions==3.2.1
django-filter==22.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
duplicity==0.8.21
executing==1.2.0
fasteners==0.14.1
filelock==3.8.2
future==0.18.2
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==5.2.0
inflection==0.5.1
ipdb==0.13.11
ipython==8.11.0
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
model-mommy==2.0.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.7
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
systemd-python==234
terminator==2.1.1
tomli==2.0.1
traitlets==5.9.0
tzdata==2022.7
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.7
virtualenv==20.17.1
wadllib==1.3.6
wcwidth==0.2.6
xdg==5
xkit==0.0.0
zipp==3.11.0
